@@ -16,8 +16,8 @@ import (
 // It is useful to mock "os/exec" functions in tests. When succeeded, it returns
 // the result produced by the test helper.
 // The test helper should:
-//     1. Use WantHelperProcess function to determine if it is being called in helper mode.
-//     2. Call fmt.Fprintln(os.Stdout, ...) to print results for the main test to collect.
+//  1. Use WantHelperProcess function to determine if it is being called in helper mode.
+//  2. Call fmt.Fprintln(os.Stdout, ...) to print results for the main test to collect.
 func Exec(helper string, envs ...string) (string, error) {
 	cmd := exec.Command(os.Args[0], "-test.run="+helper, "--")
 	cmd.Env = []string{"GO_WANT_HELPER_PROCESS=1"}
